updated: Test Straight を実装，かなり怪しいけれど動いている

+++ UserSrc/FunctionTimer.h
@@ -24,15 +24,15 @@ uint16_t FTR_GetLeftEncoderCount(void);
 void FTR_StartRightMotorTimer(void);
 void FTR_SetRightMotorDuty(float32_t duty);
 void FTR_AddRightMotorDuty(float32_t duty);
-uint16_t FTR_GetRightMotorDuty(void);
+float32_t FTR_GetRightMotorDuty(void);
 
 void FTR_StartLeftMotorTimer(void);
 void FTR_SetLeftMotorDuty(float32_t duty);
 void FTR_AddLeftMotorDuty(float32_t duty);
-uint16_t FTR_GetLeftMotorDuty(void);
+float32_t FTR_GetLeftMotorDuty(void);
 
 void FTR_StartSensorMotorTimer(void);
 void FTR_SetSensorMotorDuty(float32_t duty);
-uint16_t FTR_GetSensorMotorDuty(void);
+float32_t FTR_GetSensorMotorDuty(void);
 
 #endif /* FUNCTIONTIMER_H_ */
